tcdfvygbuh

+++ pi/gpio-pi.js
@@ -25,11 +25,11 @@ process.stdin.on("data", (data) => {
     let [cmd,...args] = data.split(' ');
     if(cmd == 'T'){
       let y = info.filter(x=>x.score==parseInt(args[0].trim()))[0];
-      y.threshold = parseInt(args[1].trim());
+      y.threshold = parseFloat(args[1].trim());
       console.log(info);
     } else if (cmd == 'N'){
       socket.emit('newgame',{name:args[0]});
-    }
+    } 
     console.log("You typed", data.trim());
 });
 
@@ -48,7 +48,6 @@ const watchHCSR04 = () => {
         const endTick = tick;
         const diff = (endTick >> 0) - (startTick >> 0); // Unsigned 32 bit arithmetic
         const temp = diff / 2 / MICROSECDONDS_PER_CM;
-        // console.log(`Sensor ${i}: ${temp}`);
         if (temp < x.threshold && !throttle) {
           console.log(`Sensor ${x.score}: ${temp} < ${x.threshold}`);
           if (new Date() - lastScore > SCORE_DELAY) {
